requirements.txt
certifi==2018.1.18
click==6.7
dominate==2.3.1
elasticsearch==6.1.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-PyMongo==0.5.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pew==1.1.2
pipenv==11.0.2
psutil==5.3.1
psycopg2==2.7.4
pymongo==3.6.1
PyMySQL==0.8.0
python-dotenv==0.8.0
SQLAlchemy==1.2.4
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
visitor==0.1.3
Werkzeug==0.14.1
wincertstore==0.2
WTForms==2.1
